feat(web): global Cmd+K palette + admin nav consolidation
Adds an app-wide command palette (chat + admin/platform) that indexes
every admin route from the canonical `admin-nav` registry plus the live
workspace settings catalog. Deep-links to `/admin/settings#setting-<key>`
scroll the row into view and pulse a highlight.

With discoverability now one keystroke away, this also collapses the
overgrown admin nav:

- Audit + Admin Action Log → /admin/audit with `?tab=actions`
- Token Usage → /admin/usage with `?tab=tokens`
- MCP settings folded into /admin/settings as a section
- Scheduler Tasks renamed "Background Jobs", moved Monitoring → Configuration
- "MFA & Sessions" relabeled "My Security"

All retired routes return server-side redirects so bookmarks survive.
Monitoring 6 → 3, Configuration 9 → 8, net −365 LOC.
