+++ packages/web/src/app/admin/token-usage/tab.tsx
@@ -1,0 +1,219 @@
+"use client";
+
+/**
+ * Token Usage rendered as a tab inside `/admin/usage`. The previous
+ * standalone `/admin/token-usage` route was folded in to deduplicate the
+ * Monitoring nav group (the two pages reported overlapping concepts —
+ * token-usage is just one slice of overall plan consumption). The old URL
+ * still resolves via a redirect on `token-usage/page.tsx`.
+ *
+ * This is a pure body component — no `<h1>`, no page-level `<div
+ * className="p-6">`. The host page (`usage/page.tsx`) provides the chrome.
+ */
+
+import dynamic from "next/dynamic";
+import { useQueryStates } from "nuqs";
+import { tokenUsageSearchParams } from "./search-params";
+import { useAdminFetch, type FetchError } from "@/ui/hooks/use-admin-fetch";
+import { TokenSummarySchema, TrendsResponseSchema, TokenUserResponseSchema } from "@/ui/lib/admin-schemas";
+import { useDarkMode } from "@/ui/hooks/use-dark-mode";
+import { formatISODate, formatNumber, parseISODate } from "@/lib/format";
+import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
+import { DatePicker } from "@/components/ui/date-picker";
+import { Button } from "@/components/ui/button";
+import { StatCard } from "@/ui/components/admin/stat-card";
+import { LoadingState } from "@/ui/components/admin/loading-state";
+import { EmptyState } from "@/ui/components/admin/empty-state";
+import { ErrorBanner } from "@/ui/components/admin/error-banner";
+import { AdminContentWrapper } from "@/ui/components/admin-content-wrapper";
+import { DataTable } from "@/components/data-table/data-table";
+import { DataTableToolbar } from "@/components/data-table/data-table-toolbar";
+import { DataTableSortList } from "@/components/data-table/data-table-sort-list";
+import { getTokenUsageColumns } from "./columns";
+import { useDataTable } from "@/hooks/use-data-table";
+import { Coins, TrendingUp, Users, MessageSquare, Search } from "lucide-react";
+import { useState } from "react";
+
+const TokenChart = dynamic(() => import("./token-chart"), { ssr: false });
+
+function buildQS(from: string, to: string): string {
+  const parts: string[] = [];
+  if (from) parts.push(`from=${encodeURIComponent(from)}`);
+  if (to) parts.push(`to=${encodeURIComponent(to)}`);
+  return parts.length > 0 ? `?${parts.join("&")}` : "";
+}
+
+function findGateError(...errors: (FetchError | null)[]): FetchError | null {
+  for (const err of errors) {
+    if (err?.status && [401, 403, 404].includes(err.status)) return err;
+  }
+  return null;
+}
+
+export function TokenUsageTab() {
+  const dark = useDarkMode();
+  const [params, setParams] = useQueryStates(tokenUsageSearchParams);
+  const [filters, setFilters] = useState({ from: params.from, to: params.to });
+
+  const qs = buildQS(params.from, params.to);
+
+  const { data: summary, loading: summaryLoading, error: summaryError } =
+    useAdminFetch(`/api/v1/admin/tokens/summary${qs}`, { schema: TokenSummarySchema, deps: [qs] });
+
+  const { data: trendsData, loading: trendsLoading, error: trendsError } =
+    useAdminFetch(`/api/v1/admin/tokens/trends${qs}`, { schema: TrendsResponseSchema, deps: [qs] });
+
+  const { data: usersData, loading: usersLoading, error: usersError } =
+    useAdminFetch(`/api/v1/admin/tokens/by-user${qs}`, { schema: TokenUserResponseSchema, deps: [qs] });
+
+  const tokenColumns = getTokenUsageColumns();
+  const tokenData = usersData?.users ?? [];
+  const { table: tokenTable } = useDataTable({
+    data: tokenData,
+    columns: tokenColumns,
+    pageCount: 1,
+    initialState: {
+      sorting: [{ id: "totalTokens", desc: true }],
+      pagination: { pageIndex: 0, pageSize: 50 },
+    },
+    getRowId: (row) => row.userId,
+  });
+
+  const gateError = findGateError(summaryError, trendsError, usersError);
+
+  function applyFilters() {
+    setParams({ from: filters.from, to: filters.to });
+  }
+
+  function clearFilters() {
+    setFilters({ from: "", to: "" });
+    setParams({ from: "", to: "" });
+  }
+
+  return (
+    <AdminContentWrapper
+      loading={false}
+      error={gateError}
+      feature="Token Usage"
+      onRetry={() => window.location.reload()}
+    >
+      <div className="space-y-6">
+        <Card className="shadow-none">
+          <CardContent className="flex flex-wrap items-end gap-3 pt-4">
+            <div className="grid gap-1">
+              <label htmlFor="from" className="text-xs text-muted-foreground">From</label>
+              <DatePicker
+                id="from"
+                value={parseISODate(filters.from)}
+                onChange={(d) => setFilters((f) => ({ ...f, from: formatISODate(d) }))}
+              />
+            </div>
+            <div className="grid gap-1">
+              <label htmlFor="to" className="text-xs text-muted-foreground">To</label>
+              <DatePicker
+                id="to"
+                value={parseISODate(filters.to)}
+                onChange={(d) => setFilters((f) => ({ ...f, to: formatISODate(d) }))}
+              />
+            </div>
+            <Button size="sm" variant="outline" className="h-8" onClick={applyFilters}>
+              <Search className="mr-1 size-3" />
+              Apply
+            </Button>
+            {(params.from || params.to) && (
+              <Button size="sm" variant="ghost" className="h-8" onClick={clearFilters}>
+                Clear
+              </Button>
+            )}
+          </CardContent>
+        </Card>
+
+        {summaryError ? (
+          <ErrorBanner message={summaryError.message} />
+        ) : summaryLoading ? (
+          <LoadingState message="Loading summary..." />
+        ) : summary ? (
+          <div className="grid gap-4 sm:grid-cols-2 lg:grid-cols-4">
+            <StatCard
+              title="Total Tokens"
+              value={formatNumber(summary.totalTokens)}
+              icon={<Coins className="size-4" />}
+            />
+            <StatCard
+              title="Prompt Tokens"
+              value={formatNumber(summary.totalPromptTokens)}
+              icon={<TrendingUp className="size-4" />}
+              description={summary.totalTokens > 0
+                ? `${((summary.totalPromptTokens / summary.totalTokens) * 100).toFixed(0)}% of total`
+                : undefined}
+            />
+            <StatCard
+              title="Completion Tokens"
+              value={formatNumber(summary.totalCompletionTokens)}
+              icon={<TrendingUp className="size-4" />}
+              description={summary.totalTokens > 0
+                ? `${((summary.totalCompletionTokens / summary.totalTokens) * 100).toFixed(0)}% of total`
+                : undefined}
+            />
+            <StatCard
+              title="Total Requests"
+              value={formatNumber(summary.totalRequests)}
+              icon={<MessageSquare className="size-4" />}
+              description={summary.totalRequests > 0 && summary.totalTokens > 0
+                ? `~${formatNumber(Math.round(summary.totalTokens / summary.totalRequests))} tokens/req`
+                : undefined}
+            />
+          </div>
+        ) : null}
+
+        <Card>
+          <CardHeader className="pb-2">
+            <CardTitle className="flex items-center gap-2 text-base">
+              <TrendingUp className="size-4" />
+              Token Usage Over Time
+            </CardTitle>
+          </CardHeader>
+          <CardContent>
+            {trendsError ? (
+              <ErrorBanner message={trendsError.message} />
+            ) : trendsLoading ? (
+              <div className="flex h-64 items-center justify-center">
+                <LoadingState message="Loading trends..." />
+              </div>
+            ) : !trendsData?.trends?.length ? (
+              <EmptyState icon={TrendingUp} title="No token usage data for this period" />
+            ) : (
+              <TokenChart data={trendsData.trends} dark={dark} />
+            )}
+          </CardContent>
+        </Card>
+
+        <Card>
+          <CardHeader className="pb-2">
+            <CardTitle className="flex items-center gap-2 text-base">
+              <Users className="size-4" />
+              Top Users by Token Consumption
+            </CardTitle>
+          </CardHeader>
+          <CardContent>
+            {usersError ? (
+              <ErrorBanner message={usersError.message} />
+            ) : usersLoading ? (
+              <div className="flex h-40 items-center justify-center">
+                <LoadingState message="Loading..." />
+              </div>
+            ) : !usersData?.users?.length ? (
+              <EmptyState icon={Users} title="No user data for this period" />
+            ) : (
+              <DataTable table={tokenTable}>
+                <DataTableToolbar table={tokenTable}>
+                  <DataTableSortList table={tokenTable} />
+                </DataTableToolbar>
+              </DataTable>
+            )}
+          </CardContent>
+        </Card>
+      </div>
+    </AdminContentWrapper>
+  );
+}
