+++ packages/web/src/ui/components/palette/index.ts
@@ -1,0 +1,3 @@
+export { GlobalCommandPalette } from "./global-command-palette";
+export type { PaletteAction, PaletteGroup, PaletteItem } from "./palette-types";
+export { PALETTE_EVENT, SHORTCUTS_EVENT } from "@/ui/components/chat/palette-events";
